pyproject.toml
[project]
name = "crawl4ai-mcp"
version = "0.1.0"
description = "MCP server for integrating web crawling and RAG into AI agents and AI coding assistants"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crawl4ai==0.6.2",
    "fastmcp==2.9.0",
    "asyncpg>=0.29.0",
    "openai>=1.71.0",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=4.1.0",
    "fastapi>=0.116.1",
    "uvicorn>=0.34.2",
    "jinja2>=3.1.6",
    "python-multipart>=0.0.20",
    "browser-cookie3>=0.19.1",
    "aiohttp>=3.8.0",
    "matplotlib>=3.10.5",
    "pandas>=2.3.1",
    "numpy>=2.2.5",
]

[tool.uv]
post-install = [
    "playwright install chromium",
    "playwright install-deps chromium"
]
